[externals,kernel,numerics] Optionally don't install csparse headers.
This patch makes it possible to not install csparse.h.

Although CSparse is inherently part of the SparseMatrix API in
numerics, it may be desirable to install Siconos without requiring
csparse.h to be present, since it is an external library.  Given that
all CSparse matrices are refered to by pointer in the numerics API,
here we replace it with an opaque struct prototype.  Testing code is
modified to include csparse.h internally, but it does not appear in
the examples.

This of course makes many functions in NumericsMatrix.h and
NumericsSparseMatrix.h unusable outside the Siconos codebase, unless
the user code has independent access to csparse.h (e.g. the same
version of CSparse already installed on the system as a separate
library.)

New option: INSTALL_EXTERNALS_HEADERS_CSPARSE, default=ON.