When getting an index array for csparse, request contiguous data.
Conversion of CSparseMatrix was broken on `armhf` architecture because the default int
type was 32 bits but allocated (during copy in coo_matrix constructor) with a stride of 8,
and the stride was not taken into account.  This commit uses PyArray_GETCONTIGUOUS to let
numpy handle the conversion to the desired type and contiguity property; it will still
point to the original memory if no conversion is necessary.